requirements.txt
streamlit
chromadb
pandas
openpyxl
ollama
pyyaml
